Add more expalanation of "Address already in use" message of coordinator. (At least one user found this too cryptic.) Add same updates to unstable branch of MTCP (size_t instead of int) that were contributed by a user for the main branch.

+++ mtcp/mtcp_restart_nolibc.c
@@ -74,11 +74,11 @@ asm (".text");
 static void readfiledescrs (void);
 static void readmemoryareas (void);
 static void readcs (char cs);
-static void readfile (void *buf, int size);
-static void mmapfile(void *buf, int size, int prot, int flags);
-static void skipfile(int size);
+static void readfile (void *buf, size_t size);
+static void mmapfile(void *buf, size_t size, int prot, int flags);
+static void skipfile(size_t size);
 static VA highest_userspace_address (VA *vdso_addr, VA *vsyscall_addr,
-				     VA * stack_end_addr);
+                                     VA * stack_end_addr);
 static int open_shared_file(char* fileName);
 // These will all go away when we use a linker to reserve space.
 static VA global_vdso_addr = 0;
@@ -503,8 +503,8 @@ static void readmemoryareas (void)
          */
 /* ANALYZE THE CONDITION FOR DOING mmapfile MORE CAREFULLY. */
         if (do_mmap_ckpt_image
-	    && mystrstr(area.name, "[vdso]")
-	    && mystrstr(area.name, "[vsyscall]"))
+            && mystrstr(area.name, "[vdso]")
+            && mystrstr(area.name, "[vsyscall]"))
           mmapfile (area.addr, area.size, area.prot | PROT_WRITE, area.flags);
         else
           readfile (area.addr, area.size);
@@ -674,9 +674,9 @@ static void readcs (char cs)
   }
 }
 
-static void readfile(void *buf, int size)
+static void readfile(void *buf, size_t size)
 {
-    int rc, ar;
+    size_t rc, ar;
     ar = 0;
 
     while(ar != size)
@@ -689,7 +689,7 @@ static void readfile(void *buf, int size)
         }
         else if(rc == 0)
         {
-            mtcp_printf("mtcp_restart_nolibc readfile: only read %d bytes instead of %d from checkpoint file\n", ar, size);
+            mtcp_printf("mtcp_restart_nolibc readfile: only read %zu bytes instead of %zu from checkpoint file\n", ar, size);
             mtcp_abort();
         }
 
@@ -697,7 +697,7 @@ static void readfile(void *buf, int size)
     }
 }
 
-static void mmapfile(void *buf, int size, int prot, int flags)
+static void mmapfile(void *buf, size_t size, int prot, int flags)
 {
     void *addr;
     int rc, ar;
@@ -708,19 +708,19 @@ static void mmapfile(void *buf, int size, int prot, int flags)
     if (addr != buf) {
         if (addr == MAP_FAILED)
             mtcp_printf("mtcp_restart_nolibc mmapfile:"
-		        " error %d reading checkpoint file\n", mtcp_sys_errno);
-	else
-	    mtcp_printf("mmapfile: Requested address %p, but got address %p\n",
-			buf, addr);
+                        " error %d reading checkpoint file\n", mtcp_sys_errno);
+        else
+            mtcp_printf("mmapfile: Requested address %p, but got address %p\n",
+                        buf, addr);
         mtcp_abort();
     }
     /* Now update mtcp_restore_cpfd so as to work the same way as readfile() */
     rc = mtcp_sys_lseek(mtcp_restore_cpfd, size, SEEK_CUR);
 }
 
-static void skipfile(int size)
+static void skipfile(size_t size)
 {
-    int rc, ar;
+    size_t rc, ar;
     ar = 0;
     char array[512];
 
@@ -734,7 +734,7 @@ static void skipfile(int size)
         }
         else if(rc == 0)
         {
-            mtcp_printf("mtcp_restart_nolibc skipfile: only skipped %d bytes instead of %d from checkpoint file\n", ar, size);
+            mtcp_printf("mtcp_restart_nolibc skipfile: only skipped %zu bytes instead of %zu from checkpoint file\n", ar, size);
             mtcp_abort();
         }
